refactor: group Query Store diagnostics files into a feature folder
The Watchdogs folder followed an implicit one-file-per-watchdog
convention. This feature had added six flat files to it, becoming ~40%
of the folder and burying the six sibling watchdogs.

Move the feature's files into Features/Watchdogs/QueryStoreDiagnostics/,
with the three log payload shapes under a Models/ subfolder, and mirror
the same structure in the unit test project. Namespaces follow the
folder layout, matching the convention used by Schema/Model,
Search/Expressions, Storage/TvpRowGeneration and Operations/Import.

All moves are pure relocations - no behaviour changes. Both projects are
SDK-style with automatic globbing, so no csproj edits are needed.
